Extracted BlibOutputWriter from MergeNodeTask (Phase A)
* Moved WriteBlibFile + WriteRetentionTimes out of MergeNodeTask into a
  new BlibOutputWriter collaborator (distinct from the low-level
  BlibWriter SQLite layer).
* Decomposed the orchestrator into a sequencer: CreateSourceFiles ->
  PrecompressSpectra -> EmitSpectrumRows -> WriteMetadata.
* Pure code motion: byte-identical blib on Stellar (modes 1 and 2).

See ai/todos/active/TODO-20260616_ospreysharp_debt_paydown_pr4.md
