+++ pwiz_tools/OspreySharp/OspreySharp.Tasks/MergeNodeTask.cs
@@ -25,8 +25,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using pwiz.OspreySharp.Core;
 using pwiz.OspreySharp.FDR;
 using pwiz.OspreySharp.IO;
@@ -326,7 +324,7 @@ namespace pwiz.OspreySharp.Tasks
             ctx.LogInfo(string.Format(
                 "[COUNT] Cross-file observations to write: {0}", nCrossFileObservations));
 
-            WriteBlibFile(config, perFileEntries, libraryById, bestByPrecursor,
+            BlibOutputWriter.Write(config, perFileEntries, libraryById, bestByPrecursor,
                 bestExpPrecursorQ, sharedBounds, entriesByPrecursor);
 
             ctx.LogInfo(string.Format("Wrote {0} spectra to {1}",
@@ -526,238 +524,6 @@ namespace pwiz.OspreySharp.Tasks
                 }
             }
             return entriesByPrecursor;
-        }
-
-        // Per-observation RetentionTimes rows — one row for EVERY run where this
-        // precursor was detected. retentionTime (drives Skyline ID-line display)
-        // is populated iff the run passes run-level FDR, OR (fallback) no run
-        // passes and this is the best run by lowest run_qvalue. Cross-charge
-        // shared boundaries applied. Mirrors Rust pipeline.rs:6191-6243.
-        private static void WriteRetentionTimes(
-            BlibWriter writer, long refId, string fileName,
-            List<KeyValuePair<string, FdrEntry>> observations,
-            Dictionary<string, long> sourceFileIds,
-            Dictionary<(string, string), double[]> sharedBounds,
-            double fdrThreshold)
-        {
-            if (observations == null)
-                return;
-            // Compute the fallback ID-line file: if NO run passes run-level FDR,
-            // the run with the lowest run_qvalue gets the ID line so every blib
-            // RefSpectra has at least one ID line.
-            bool anyPassesRunFdr = false;
-            string bestRunFile = null;
-            double bestRunQ = double.MaxValue;
-            foreach (var obs in observations)
-            {
-                double rq = obs.Value.EffectiveRunQvalue(FdrLevel.Both);
-                if (rq <= fdrThreshold)
-                    anyPassesRunFdr = true;
-                if (rq < bestRunQ)
-                {
-                    bestRunQ = rq;
-                    bestRunFile = obs.Key;
-                }
-            }
-
-            foreach (var obs in observations)
-            {
-                long srcId = sourceFileIds[obs.Key];
-                var fileEntry = obs.Value;
-                double runQ = fileEntry.EffectiveRunQvalue(FdrLevel.Both);
-                bool passesFdr = runQ <= fdrThreshold;
-                bool showIdLine = passesFdr ||
-                    (!anyPassesRunFdr && obs.Key == bestRunFile);
-                bool isBest = obs.Key == fileName;
-
-                var runSharedKey = (fileEntry.ModifiedSequence, obs.Key);
-                double runApex = fileEntry.ApexRt;
-                double runStart = fileEntry.StartRt;
-                double runEnd = fileEntry.EndRt;
-                double[] runShared;
-                if (sharedBounds.TryGetValue(runSharedKey, out runShared))
-                {
-                    runApex = runShared[0];
-                    runStart = runShared[1];
-                    runEnd = runShared[2];
-                }
-
-                double? rtForIdLine = null;
-                if (showIdLine)
-                    rtForIdLine = runApex;
-                writer.AddRetentionTime(
-                    refId, srcId,
-                    rtForIdLine,
-                    runStart,
-                    runEnd,
-                    runQ,
-                    isBest);
-            }
-        }
-
-        // Write the .blib: source file IDs, parallel zlib pre-compress, then the
-        // sequential per-best-precursor RefSpectra + modifications + protein +
-        // RetentionTimes + Osprey extension-table emission, metadata, finalize.
-        private static void WriteBlibFile(
-            OspreyConfig config,
-            List<KeyValuePair<string, List<FdrEntry>>> perFileEntries,
-            IReadOnlyDictionary<uint, LibraryEntry> libraryById,
-            Dictionary<(string, byte), KeyValuePair<string, FdrEntry>> bestByPrecursor,
-            Dictionary<(string, byte), double> bestExpPrecursorQ,
-            Dictionary<(string, string), double[]> sharedBounds,
-            Dictionary<(string, byte), List<KeyValuePair<string, FdrEntry>>> entriesByPrecursor)
-        {
-            double fdrThreshold = config.RunFdr; // run-level threshold for ID-line semantics
-            using (var writer = new BlibWriter(config.OutputBlib))
-            {
-                writer.BeginBatch();
-
-                // Pre-create source file IDs once. SpectrumSourceFiles.idFileName
-                // carries the library filename (Skyline expects this — Rust
-                // pipeline.rs:6110 + blib.rs:435). The library file is the "ID
-                // source"; the mzML file is the spectrum source.
-                string libraryIdName = Path.GetFileName(config.LibrarySource.Path);
-                var sourceFileIds = new Dictionary<string, long>();
-                foreach (var kvp in perFileEntries)
-                {
-                    sourceFileIds[kvp.Key] = writer.AddSourceFile(
-                        kvp.Key + ".mzML", libraryIdName, fdrThreshold);
-                }
-
-                // Parallel pre-compress pass. Per-spectrum zlib dominates the blib
-                // write wall; pre-compute (mzBlob, intBlob, numPeaks) for every
-                // entry in parallel, then drive AddSpectrumPrecompressed in
-                // iteration order so RefSpectra row IDs stay deterministic.
-                var blibEntries = bestByPrecursor.Values.ToList();
-                int blibN = blibEntries.Count;
-                var blibMzBlobs = new byte[blibN][];
-                var blibIntBlobs = new byte[blibN][];
-                var blibNumPeaks = new int[blibN];
-                Parallel.For(0, blibN,
-                    new ParallelOptions { MaxDegreeOfParallelism = config.NThreads },
-                    i =>
-                    {
-                        var entry = blibEntries[i].Value;
-                        LibraryEntry libEntryP;
-                        if (!libraryById.TryGetValue(entry.EntryId, out libEntryP))
-                            return;
-                        int nFrags = libEntryP.Fragments.Count;
-                        var mzsP = new double[nFrags];
-                        var intsP = new float[nFrags];
-                        for (int j = 0; j < nFrags; j++)
-                        {
-                            mzsP[j] = libEntryP.Fragments[j].Mz;
-                            intsP[j] = libEntryP.Fragments[j].RelativeIntensity;
-                        }
-                        blibMzBlobs[i] = BlibWriter.CompressMzs(mzsP);
-                        blibIntBlobs[i] = BlibWriter.CompressIntensities(intsP);
-                        blibNumPeaks[i] = nFrags;
-                    });
-
-                for (int blibIdx = 0; blibIdx < blibN; blibIdx++)
-                {
-                    var kvp = blibEntries[blibIdx];
-                    string fileName = kvp.Key;
-                    var entry = kvp.Value;
-
-                    LibraryEntry libEntry;
-                    if (!libraryById.TryGetValue(entry.EntryId, out libEntry))
-                        continue;
-
-                    long fileId = sourceFileIds[fileName];
-
-                    byte[] mzBlobPre = blibMzBlobs[blibIdx];
-                    byte[] intBlobPre = blibIntBlobs[blibIdx];
-                    int numPeaksPre = blibNumPeaks[blibIdx];
-
-                    // RefSpectra.score is the EXPERIMENT-PRECURSOR q-value (min
-                    // across all observations of this (modseq, charge)). Mirrors
-                    // Rust pipeline.rs:4670-4683 / 4795. Same value feeds
-                    // OspreyExperimentScores.ExperimentQValue below.
-                    var lookupKey = (entry.ModifiedSequence, entry.Charge);
-                    double scoreQvalue;
-                    if (!bestExpPrecursorQ.TryGetValue(lookupKey, out scoreQvalue))
-                        scoreQvalue = entry.ExperimentPrecursorQvalue;
-
-                    // nRunsDetected -> RefSpectra.copies (Rust pipeline.rs:6179
-                    // passes n_runs_detected = group.len()). Reused by
-                    // OspreyExperimentScores below.
-                    List<KeyValuePair<string, FdrEntry>> observations;
-                    int nRunsDetected = 1;
-                    if (entriesByPrecursor.TryGetValue(lookupKey, out observations) &&
-                        observations.Count > 0)
-                    {
-                        nRunsDetected = observations.Count;
-                    }
-
-                    // Shared peak boundaries when the peptide is detected at
-                    // multiple charges in this file (Rust pipeline.rs:6160-6164).
-                    var sharedKey = (entry.ModifiedSequence, fileName);
-                    double sharedApex = entry.ApexRt;
-                    double sharedStart = entry.StartRt;
-                    double sharedEnd = entry.EndRt;
-                    double[] sharedVals;
-                    if (sharedBounds.TryGetValue(sharedKey, out sharedVals))
-                    {
-                        sharedApex = sharedVals[0];
-                        sharedStart = sharedVals[1];
-                        sharedEnd = sharedVals[2];
-                    }
-
-                    long refId = writer.AddSpectrumPrecompressed(
-                        libEntry.Sequence,
-                        libEntry.ModifiedSequence,
-                        libEntry.PrecursorMz,
-                        libEntry.Charge,
-                        sharedApex,
-                        sharedStart,
-                        sharedEnd,
-                        mzBlobPre, intBlobPre, numPeaksPre,
-                        scoreQvalue, fileId, nRunsDetected, 0.0);
-
-                    // Add modifications
-                    if (libEntry.Modifications != null && libEntry.Modifications.Count > 0)
-                        writer.AddModifications(refId, libEntry.Modifications);
-
-                    // Add protein mappings
-                    if (libEntry.ProteinIds != null && libEntry.ProteinIds.Count > 0)
-                        writer.AddProteinMapping(refId, libEntry.ProteinIds);
-
-                    WriteRetentionTimes(writer, refId, fileName, observations,
-                        sourceFileIds, sharedBounds, fdrThreshold);
-
-                    // Osprey extension tables — one row per RefSpectra each,
-                    // mirroring Rust pipeline.rs:6255-6272. Best-run-only for
-                    // OspreyPeakBoundaries + OspreyRunScores; experiment-level for
-                    // OspreyExperimentScores. The 0.0 fields are the same "not yet
-                    // plumbed through Stage 7 plan entries" placeholders Rust writes.
-                    writer.AddPeakBoundaries(refId, fileName,
-                        sharedStart, sharedEnd, sharedApex,
-                        0.0, // ApexIntensity — matches Rust's apex_coefficient placeholder
-                        entry.BoundsArea);
-                    writer.AddRunScores(refId, fileName,
-                        entry.EffectiveRunQvalue(FdrLevel.Both),
-                        0.0, // DiscriminantScore — matches Rust's dot_product placeholder
-                        0.0); // PosteriorErrorProb — matches Rust's PEP placeholder
-                    writer.AddExperimentScores(refId,
-                        scoreQvalue, // Same value as RefSpectra.score
-                        nRunsDetected,
-                        perFileEntries.Count);
-                }
-
-                writer.Commit();
-
-                // Add metadata. OspreyMetadata key set must match Rust's
-                // write_blib_from_plan (pipeline.rs:6078-6081) byte-for-byte.
-                writer.AddMetadata(@"osprey_version", OspreyVersion.Current);
-                writer.AddMetadata(@"search_mode", @"coelution");
-                writer.AddMetadata(@"run_fdr",
-                    config.RunFdr.ToString(System.Globalization.CultureInfo.InvariantCulture));
-                writer.AddMetadata(@"experiment_fdr",
-                    config.ExperimentFdr.ToString(System.Globalization.CultureInfo.InvariantCulture));
-
-                writer.FinalizeDatabase();
-            }
         }
     }
 }
